pyproject.toml
[project]
name = "byte-me"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apispec-webframeworks>=1.2.0",
    "flask>=3.1.1",
    "flask-bcrypt>=1.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-marshmallow>=1.3.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
]

[tool.pytest.ini_options]
addopts = "-ra -v"
testpaths = [
    "tests",
]
